Refactor and new widgets / flows
